tests/psoc6/spi: Refactoring SPI tests. WIP.
